requirements.txt
django-model-utils>=2.0
requests==2.21.0
